Style: Reformat all test code
