+++ src/test/java/logic/ItemTest.java
@@ -5,13 +5,13 @@ import org.junit.Test;
 
 public class ItemTest {
 
-    Item item = new Item(20201125);
+  Item item = new Item(20201125);
 
-    @Test
-    public void testGetExpiration_Date() {
-        item.setExpirationDate(20211026);
-        int ExpectedResult = 20211026;
-        int ActualResult = item.getExpirationDate();
-        Assert.assertEquals(ExpectedResult, ActualResult);
-    }
+  @Test
+  public void testGetExpiration_Date() {
+    item.setExpirationDate(20211026);
+    int ExpectedResult = 20211026;
+    int ActualResult = item.getExpirationDate();
+    Assert.assertEquals(ExpectedResult, ActualResult);
+  }
 }
